+++ features/TubenessProcessor.java
@@ -4,7 +4,7 @@ package features;
 
 public class TubenessProcessor extends HessianEvalueProcessor {
 
-	public double measureFromEvalues2D( double [] evalues ) {
+	public float measureFromEvalues2D( float [] evalues ) {
 
 		/* If either of the two principle eigenvalues is
 		   positive then the curvature is in the wrong
@@ -14,10 +14,10 @@ public class TubenessProcessor extends HessianEvalueProcessor {
 		if ((evalues[1] >= 0) || (evalues[0] >= 0))
 			return 0;
 		else
-			return Math.abs(evalues[1]);
+			return (float)Math.abs(evalues[1]);
 	}
 
-	public double measureFromEvalues3D( double [] evalues ) {
+	public float measureFromEvalues3D( float [] evalues ) {
 
 		/* If either of the two principle eigenvalues is
 		   positive then the curvature is in the wrong
@@ -27,6 +27,6 @@ public class TubenessProcessor extends HessianEvalueProcessor {
 		if ((evalues[1] >= 0) || (evalues[2] >= 0))
 			return 0;
 		else
-			return Math.sqrt(evalues[2] * evalues[1]);
+			return (float)Math.sqrt(evalues[2] * evalues[1]);
 	}
 }
